requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.3
openpyxl==3.1.2
msal==1.26.0
O365==2.0.34 